Moving the exceptions to src/main/java
