+++ src/test/java/com/adobe/campaign/tests/logparser/core/SDKTests.java
@@ -8,7 +8,11 @@
  */
 package com.adobe.campaign.tests.logparser.core;
 
-import com.adobe.campaign.tests.logparser.data.*;
+
+import com.adobe.campaign.tests.logparser.data.SDKCaseBadDefConstructor;
+import com.adobe.campaign.tests.logparser.data.SDKCasePrivateDefConstructor;
+import com.adobe.campaign.tests.logparser.exceptions.SDKCaseNoDefConstructor;
+import com.adobe.campaign.tests.logparser.data.SDKCaseSTD;
 import com.adobe.campaign.tests.logparser.exceptions.LogDataExportToFileException;
 import com.adobe.campaign.tests.logparser.exceptions.LogParserSDKDefinitionException;
 import com.adobe.campaign.tests.logparser.exceptions.StringParseException;
@@ -17,7 +21,6 @@ import org.testng.annotations.Test;
 
 import java.io.File;
 import java.util.Arrays;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -105,8 +108,8 @@ public class SDKTests {
 
         l_pDefinition.setStorePathFrom((new File("src")).getAbsolutePath());
 
-        LogData<ACCLogs> l_entries = LogDataFactory.generateLogData(l_file, "*.log", l_pDefinition,
-                ACCLogs.class);
+        LogData<SDKCaseSTD> l_entries = LogDataFactory.generateLogData(l_file, "*.log", l_pDefinition,
+                SDKCaseSTD.class);
 
         assertThat("We should have a correct number of errors", l_entries.getEntries().size(), is(equalTo(14)));
         AssertLogData.assertLogContains(l_entries, "errorMessage",
@@ -116,7 +119,7 @@ public class SDKTests {
         assertThat("We should have a file name", selectedItem.getFileName(),
                 is(equalTo("useCase1.log")));
 
-        l_entries.exportLogDataToCSV();
+        //l_entries.exportLogDataToCSV();
 
     }
